requirements.txt
alembic==1.18.4
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
dnspython==2.8.0
email-validator==2.3.0
Flask==3.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
pillow==12.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
six==1.17.0
SQLAlchemy==2.0.23
tomli==2.4.1
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
Werkzeug==3.0.1
WTForms==3.2.1
